Added tests with `DROP TABLE` and split into api, command line, and others
