+++ src/test/resources/migration/integ_outoforder3/V1_2_0__First_delete_temp.cql
@@ -1,0 +1,1 @@
+DROP TABLE IF EXISTS test2;
